fix(CalculatorIO):Long 값 int로 받아오는 현상 해결.

+++ src/main/java/calculator/View/CalculatorIO.java
@@ -20,7 +20,7 @@ public class CalculatorIO {
 
     }
 
-    public void Output(int result) {
+    public void Output(Long result) {
         System.out.printf(Output_String, result);
     }
